Fixed the duplicate link compliant
- Fixed the duplicate link compliant in preparation for submission in WordPress Plugin Repository

"The plugin has duplicate plugin and author URLs. A plugin URL is a page/site that provides details about this specific plugin. An author URL is a page/site that provides information about the author of the plugin. You are not required to provide both, so pick the one that best applies to your URL."